wip: move Facet fields to facets module
