+++ umap/static/umap/js/modules/facets.js
@@ -3,6 +3,7 @@ import { translate } from './i18n.js'
 import { Form } from './form/builder.js'
 import * as Utils from './utils.js'
 import Orderable from './orderable.js'
+import { Fields } from './form/fields.js'
 
 const WIDGETS = ['checkbox', 'radio', 'minmax']
 
@@ -337,5 +338,197 @@ export default class Facets {
     const form = new FacetsForm(this, this.build())
     container.appendChild(form.build())
     return form
+  }
+}
+
+Fields.FacetSearchBase = class extends Fields.Base {
+  buildLabel() {}
+}
+
+Fields.FacetSearchChoices = class extends Fields.FacetSearchBase {
+  getTemplate() {
+    return `
+      <fieldset class="umap-facet">
+        <legend data-ref=label>${this.properties.label}</legend>
+        <ul data-ref=ul></ul>
+      </fieldset>
+      `
+  }
+
+  build() {
+    this.type = this.properties.criteria.widget || 'checkbox'
+
+    const choices = this.properties.criteria.choices
+    choices.sort()
+    choices.forEach((value) => this.buildLi(value))
+    super.build()
+  }
+
+  buildLi(value) {
+    const name = `${this.type}_${this.name}`
+    const [li, { input, label }] = Utils.loadTemplateWithRefs(`
+      <li>
+        <label>
+          <input type="${this.type}" name="${name}" data-ref=input />
+          <span data-ref=label></span>
+        </label>
+      </li>
+    `)
+    label.textContent = value
+    input.checked = this.get()?.choices?.includes(value)
+    input.dataset.value = value
+    input.addEventListener('change', () => this.sync())
+    this.elements.ul.appendChild(li)
+  }
+
+  toJS() {
+    return {
+      type: this.type,
+      choices: [...this.elements.ul.querySelectorAll('input:checked')].map(
+        (i) => i.dataset.value
+      ),
+    }
+  }
+}
+
+Fields.MinMaxBase = class extends Fields.FacetSearchBase {
+  getInputType(type) {
+    return type
+  }
+
+  getLabels() {
+    return [translate('Min'), translate('Max')]
+  }
+
+  prepareForHTML(value) {
+    return value.valueOf()
+  }
+
+  getTemplate() {
+    const [minLabel, maxLabel] = this.getLabels()
+    const { min, max, widget } = this.properties.criteria
+    this.type = widget
+    const inputType = this.getInputType(this.type)
+    const minHTML = this.prepareForHTML(min)
+    const maxHTML = this.prepareForHTML(max)
+    return `
+      <fieldset class="umap-facet">
+        <legend>${this.properties.label}</legend>
+        <label>${minLabel}<input min="${minHTML}" max="${maxHTML}" step=any type="${inputType}" data-ref=minInput /></label>
+        <label>${maxLabel}<input min="${minHTML}" max="${maxHTML}" step=any type="${inputType}" data-ref=maxInput /></label>
+      </fieldset>
+    `
+  }
+
+  build() {
+    this.minInput = this.elements.minInput
+    this.maxInput = this.elements.maxInput
+    const { min, max, type } = this.properties.criteria
+    const { min: modifiedMin, max: modifiedMax } = this.get() || {}
+
+    const currentMin = modifiedMin !== undefined ? modifiedMin : min
+    const currentMax = modifiedMax !== undefined ? modifiedMax : max
+    if (min != null) {
+      // The value stored using setAttribute is not modified by
+      // user input, and will be used as initial value when calling
+      // form.reset(), and can also be retrieve later on by using
+      // getAttributing, to compare with current value and know
+      // if this value has been modified by the user
+      // https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/reset
+      this.minInput.setAttribute('value', this.prepareForHTML(min))
+      this.minInput.value = this.prepareForHTML(currentMin)
+    }
+
+    if (max != null) {
+      // Cf comment above about setAttribute vs value
+      this.maxInput.setAttribute('value', this.prepareForHTML(max))
+      this.maxInput.value = this.prepareForHTML(currentMax)
+    }
+    this.toggleStatus()
+
+    this.minInput.addEventListener('change', () => this.sync())
+    this.maxInput.addEventListener('change', () => this.sync())
+    super.build()
+  }
+
+  toggleStatus() {
+    this.minInput.dataset.modified = this.isMinModified()
+    this.maxInput.dataset.modified = this.isMaxModified()
+  }
+
+  sync() {
+    super.sync()
+    this.toggleStatus()
+  }
+
+  isMinModified() {
+    const default_ = this.minInput.getAttribute('value')
+    const current = this.minInput.value
+    return current !== default_
+  }
+
+  isMaxModified() {
+    const default_ = this.maxInput.getAttribute('value')
+    const current = this.maxInput.value
+    return current !== default_
+  }
+
+  toJS() {
+    const opts = {
+      type: this.type,
+    }
+    if (this.minInput.value !== '' && this.isMinModified()) {
+      opts.min = this.prepareForJS(this.minInput.value)
+    }
+    if (this.maxInput.value !== '' && this.isMaxModified()) {
+      opts.max = this.prepareForJS(this.maxInput.value)
+    }
+    return opts
+  }
+}
+
+Fields.FacetSearchNumber = class extends Fields.MinMaxBase {
+  getInputType(type) {
+    return 'number'
+  }
+
+  prepareForJS(value) {
+    return new Number(value)
+  }
+}
+
+Fields.FacetSearchDate = class extends Fields.MinMaxBase {
+  getInputType(type) {
+    return 'date'
+  }
+
+  prepareForJS(value) {
+    return new Date(value)
+  }
+
+  toLocaleDateTime(dt) {
+    return new Date(dt.valueOf() - dt.getTimezoneOffset() * 60000)
+  }
+
+  prepareForHTML(value) {
+    // Value must be in local time
+    if (!value || isNaN(value)) return
+    return this.toLocaleDateTime(value).toISOString().substr(0, 10)
+  }
+
+  getLabels() {
+    return [translate('From'), translate('Until')]
+  }
+}
+
+Fields.FacetSearchDateTime = class extends Fields.FacetSearchDate {
+  getInputType() {
+    return 'datetime-local'
+  }
+
+  prepareForHTML(value) {
+    // Value must be in local time
+    if (Number.isNaN(value)) return
+    return this.toLocaleDateTime(value).toISOString().slice(0, -1)
   }
 }
