+++ umap/static/umap/js/modules/form/builder.js
@@ -1,7 +1,7 @@
 import { translate } from '../i18n.js'
 import { SCHEMA } from '../schema.js'
 import * as Utils from '../utils.js'
-import getClass from './fields.js'
+import { getClass } from './fields.js'
 
 export class Form extends Utils.WithEvents {
   constructor(obj, fields, properties) {
